pyproject.toml
[project]
name = "gptemotion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diskcache>=5.6.3",
    "fastexcel>=0.12.0",
    "openai>=1.56.2",
    "polars>=1.16.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "xlsxwriter>=3.2.0",
]
